requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
telepot~=12.7
Flask~=3.1.0
bs4~=0.0.2
beautifulsoup4~=4.13.3